add user query param to tracked/ endpoint
